changed played time data type from float to double

+++ components/esm/savedgame.hpp
@@ -29,7 +29,7 @@ namespace ESM
         std::string mPlayerClass;
         std::string mPlayerCell;
         TimeStamp mInGameTime;
-        float mTimePlayed;
+        double mTimePlayed;
 
         /// \todo add field for screenshot
